Use generic Dictionary implementation
The old code was hard-coded to use the `PrefixedStringDictionary` in
many places. This made testing with other `Dictionary` implementations
quite cumbersome. To address this, everything is now generic in the
`Dictionary` trait, und a feature
flag (`no-prefixed-string-dictionary`) is used to change the default
for the CLI application from `PrefixedStringDictionary` to
`StringDictionary`.