Drop `operations` bench
The `operations` bench is no longer useful, yet pulls in polars, a
huge dependency. Back when we did not yet have a full reasoner
implemented, verifying correctness of our joins by comparing with the
polars implementation was useful. This is now covered by, among other
things, our end-to-end tests. Polars is not only a huge dependency
taking a lot of time to compile, but also breaks its API on minor
version bumps (which is not technically a semver violation, since it's
still 0.x, but requires changes to the bench code). Furthermore,
the bench uses internal Nemo APIs and thus also frequently requires
changes, which is often only discovered in CI, since the bench takes
much too long to run regularly. Lastly, it's not pratically useable
for external people, since it depends on test data that is not freely
redistributable.